src/backend/requirements.txt
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
autoflake==2.2.1
bcrypt==4.1.1
blinker==1.7.0
cachelib==0.10.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
Deprecated==1.2.14
dnspython==2.4.2
fastapi==0.110.0
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Mail==0.9.1
Flask-PyMongo==2.3.0
Flask-Session==0.5.0
h11==0.14.0
idna==3.4
importlib-metadata==6.9.0
importlib-resources==6.1.1
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.3
mpmath==1.3.0
mysql-connector-python==8.3.0
numpy==1.26.3
opencv-python==4.9.0.80
packaging==23.2
pycparser==2.21
pydantic==2.6.3
pydantic-settings==2.2.1
pydantic_core==2.16.3
pyflakes==3.1.0
pymongo==4.6.1
python-dotenv==1.0.0
pytz==2024.1
redis==4.2.0
referencing==0.33.0
requests==2.31.0
rpds-py==0.17.1
sniffio==1.3.1
SQLAlchemy==2.0.28
sqlmodel==0.0.16
starlette==0.36.3
sympy==1.12
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.27.1
Werkzeug==2.3.8
wrapt==1.16.0
zipp==3.17.0
